fix CI

+++ src/bfs_dfs_backtracking/dijkstra.rs
@@ -36,7 +36,7 @@ fn network_delay_time(times: Vec<Vec<i32>>, n: i32, k: i32) -> i32 {
         }
     }
 
-    if node_distance_to_k.len() == n as usize {
+    if node_distance_to_k.len() == n {
         // 从节点 k 出发能传播到所有的节点
         node_distance_to_k.into_values().max().unwrap() as i32
     } else {
